+++ packages/connectors/hull-hubspot/server/lib/sync-agent/contact-meta-group.js
@@ -1,0 +1,17 @@
+const contactMetaGroup = {
+  name: "contactmeta",
+  displayName: "Contact Meta",
+  properties: [
+    {
+      name: "contact_meta.merged-vids",
+      label: "Merged Vids",
+      type: "string",
+      fieldType: "text",
+      readOnlyValue: true
+    }
+  ]
+};
+
+module.exports = {
+  contactMetaGroup
+};